+++ tests/integration_tests/test_client.py
@@ -341,3 +341,40 @@ def test_column_rename_with_bad_option(test_config: TestConfig):
             password=test_config.password,
             rename_response_column="not_an_option",
         )
+
+
+def test_role_setting_works(test_client: Client, test_config: TestConfig):
+    role_limited = 'limit_rows_role'
+    user_limited = 'limit_rows_user'
+
+    test_client.command(f'CREATE ROLE IF NOT EXISTS {role_limited}')
+    test_client.command(f'CREATE USER IF NOT EXISTS {user_limited} IDENTIFIED WITH no_password')
+    test_client.command(f'GRANT SELECT ON system.numbers TO {user_limited}')
+    test_client.command(f'GRANT {role_limited} TO {user_limited}')
+    test_client.command(f'SET DEFAULT ROLE NONE TO {user_limited}')
+
+    client = create_client(
+        host=test_client.server_host_name,
+        port=test_client.url.rsplit(':', 1)[-1].split('/')[0],
+        username=user_limited,
+        password='',
+    )
+
+    # the default should not have the role
+    res = client.query('SELECT currentRoles()')
+    assert res.result_rows == [([],)]
+
+    # passing it as a per-query setting should work
+    res = client.query('SELECT currentRoles()', settings={'role': role_limited})
+    assert res.result_rows == [([role_limited],)]
+
+    # passing it as a per-client setting should work
+    role_client = create_client(
+        host=test_client.server_host_name,
+        port=test_client.url.rsplit(':', 1)[-1].split('/')[0],
+        username=user_limited,
+        password='',
+        settings={'role': role_limited},
+    )
+    res = role_client.query('SELECT currentRoles()')
+    assert res.result_rows == [([role_limited],)]