+++ CHANGELOG.md
@@ -61,6 +61,7 @@ The supported method of passing ClickHouse server settings is to prefix such arg
 - Changed `AsyncClient.settings` typing to `Optional[Dict[str, Any]]` to accept None inputs.
 - Added more robust error handling and tests. Closes [#508](https://github.com/ClickHouse/clickhouse-connect/issues/508)
 - Replace the use of deprecated `datetime.utcfromtimestamp`
+- Allow passing `role` as a field in the `settings` keyword argument to set a role for a specific query
 
 ### Bug Fixes
 - Fixed an AttributeError on `http.client` when importing `clickhouse_connect` under certain circumstances